requirements.txt
torch==0.4.1
torchvision==0.2.1
tensorflow-gpu==1.12.0
easydict
pathlib
requests
nltk
pandas
scikit-image
pyyaml
numpy
pillow
scipy
transformers